opentelemetry-exporter-sumologic moved from lib/ to src/
