+++ src/opentelemetry-exporter-sumologic/index.ts
