+++ src/opentelemetry-exporter-sumologic/transform.ts
