+++ src/opentelemetry-exporter-sumologic/types.ts
