+++ src/opentelemetry-exporter-sumologic/utils.ts
