+++ tsconfig.json
@@ -14,9 +14,6 @@
       "@opentelemetry/*": [
         "./opentelemetry-js/packages/opentelemetry-*",
         "./opentelemetry-js-contrib/plugins/web/opentelemetry-*"
-      ],
-      "opentelemetry-exporter-sumologic": [
-        "./lib/opentelemetry-exporter-sumologic"
       ]
     }
   }
